Updated processing steps, scrapers and application widgets
