+++ src/phdfinder/scraping/scrapers/scraper_CH_ethz.py
@@ -1,12 +1,11 @@
 import datetime
-import time
-
+import os
 import bs4
 import utils.scraping_utils as su
-import os
 from selenium.webdriver.common.by import By
-from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support.ui import WebDriverWait
+
 
 def select_phd_jobs(driver_page):
     try:
