+++ src/phdfinder/scraping/scrapers/utils/scraping_utils.py
@@ -2,14 +2,12 @@ import inspect
 import os
 import time
 
-import pandas as pd
-from selenium import webdriver
-from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.common.by import By
-from selenium.webdriver.chrome.options import Options
-from webdriver_manager.chrome import ChromeDriverManager
-from bs4 import BeautifulSoup
 import polars as pl
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.chrome.service import Service
+from webdriver_manager.chrome import ChromeDriverManager
+
 
 def get_page_through_selenium(page, ad_element, return_driver=False):
     """
